envoy: cleanup xDS configuration
